cloudinary

+++ AgriConnectBack-ProduitBack/src/main/java/com/example/usermanagementbackend/repository/ProduitRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProduitRepository extends JpaRepository<Produit, Integer> {
+public interface ProduitRepository extends JpaRepository<Produit, Long> { // Changement de Integer à Long
     Page<Produit> findByNomContaining(String nom, Pageable pageable);
     Page<Produit> findByFournisseurContaining(String fournisseur, Pageable pageable);
     Page<Produit> findByPrixBetween(double min, double max, Pageable pageable);
